final commit

+++ src/migrations/20250107202211_create_user_table.js
@@ -30,7 +30,7 @@ export function up(knex) {
       .unsigned()
       .nullable();
     table.integer('phone_number').unsigned();
-    table.integer('nic').unsigned();
+    table.bigInteger('nic').unsigned();
     table.integer('managing_org').unsigned();
     table.string('blockchain_key').nullable();
